refactor: Money, Wallet 클래스 분리

+++ src/test/java/step2/ReceiptTest.java
@@ -13,7 +13,7 @@ class ReceiptTest {
     void constructor() {
         Receipt receipt = new Receipt(0, Arrays.asList(new Ticket(Arrays.asList(1, 2, 3, 4, 5, 6))));
 
-        assertThat(receipt.getChange()).isEqualTo(0);
+        assertThat(receipt.getChange().toInt()).isEqualTo(0);
         assertThat(receipt.getTickets().size()).isEqualTo(1);
         assertThat(receipt.getTickets()).isEqualTo(Arrays.asList(new Ticket(Arrays.asList(1, 2, 3, 4, 5, 6))));
     }
